CAdvancedTasksPrintNumbers

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,7 +6,7 @@
 void print_number(int n)
 {
 	char lastDigit;
-	long reversed;
+	int reversed;
 
 if (n < 0)
 {
